requirements.txt
websocket-client~=1.6.1
prometheus-client~=0.17.1